header and side panel setup and moved to header.jsx
